Finished functional of the admin

+++ src/Controller/ProfileController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Entity\MainCategory;
 use App\Entity\Product;
 use App\Entity\ProductImage;
 use App\Form\DeleteImageForm;
@@ -40,11 +41,6 @@ class ProfileController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()){
             $newProduct = $form->getData();
 
-            echo '<pre>';
-            return new Response(var_dump($newProduct));
-            echo '</pre>';
-            die;
-
             $addProduct->addItems($newProduct);
 
             return $this->render('profile/profile.html.twig', [
@@ -81,15 +77,26 @@ class ProfileController extends AbstractController
         $em = $this->getDoctrine()->getManager();
         $em->remove($productInfo);
 
-        foreach ($imagesProducts as $image){
-            $em->remove($image);
-            \unlink($pathName->getUploadImagePath().$image->getImagePath());
+        if (!empty($imagesProducts[0])) {
+            foreach ($imagesProducts as $image) {
+                $em->remove($image);
+                \unlink($pathName->getUploadImagePath() . $image->getImagePath());
+            }
         }
         $em->flush();
         return $this->redirectToRoute('homepage');
 
     }
 
+    /**
+     * editProduct method edit Product entity and add new files
+     *
+     * @param int $id
+     * @param Request $request
+     * @param PathName $path
+     * @param AddProduct $editProduct
+     * @return Response
+     */
     public function editProduct(int $id, Request $request, PathName $path, AddProduct $editProduct): Response
     {
         $product = $this->getDoctrine()
@@ -97,14 +104,18 @@ class ProfileController extends AbstractController
             ->previewFindById($id)
         ;
 
+        if (empty($product)){
+            throw $this->createNotFoundException('Product with ID: '.$id.' not found!');
+        }
+
+
         $productInfo = \array_shift($product);
         $productImages = $product;
 
         $newEditProduct = new ProductModel();
-
         $newEditProduct->setId($id);
         $newEditProduct->setName($productInfo->getName());
-        $newEditProduct->setTypeId($productInfo->getTypeId());
+        $newEditProduct->setTypeId($productInfo->getName());
         $newEditProduct->setPrice($productInfo->getPrice());
         $newEditProduct->setComment($productInfo->getComment());
 
@@ -115,21 +126,19 @@ class ProfileController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()){
             $newProduct = $form->getData();
 
-           // $editProduct->updateItem($newProduct, $id);
+            $editProduct->updateItem($newProduct, $id);
 
-
-            echo '<pre>';
-            return new Response(var_dump($newProduct));
-            echo '</pre>';
-            die;
-
-
-            return $this->render('profile/edit.html.twig', [
-                'message' => 'Congratulation your product is edited!',
-                'product' => $productInfo,
-                'productImages' => $path->getNameFile($productImages)
+            return $this->redirectToRoute('profile_edit', [
+                'id' => $id
             ]);
+        }
 
+        if (empty($productImages[0])){
+            return $this->render('profile/edit.html.twig',[
+                'messageImage' => 'Your product does not have images!',
+                'editProduct' => $form->createView(),
+                'product' => $productInfo,
+            ]);
         }
 
         return $this->render('profile/edit.html.twig', [
@@ -140,70 +149,29 @@ class ProfileController extends AbstractController
 
     }
 
+    /**
+     * deleteProductImage method deletes name of image from db and delete file from directory
+     *
+     * @param $id
+     * @param PathName $pathName
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     */
     public function deleteProductImage($id, PathName $pathName)
     {
         $productImage = $this->getDoctrine()->getRepository(ProductImage::class)
             ->find($id);
 
-        if (empty($productImage)){
-            throw $this->createNotFoundException('Product Image with ID: '.$id.' not found!');
-        }
-
         $em = $this->getDoctrine()->getManager();
         $em->remove($productImage);
         $em->flush();
-        \unlink($pathName->getUploadImagePath().$productImage->getImagePath());
+
+
+        if (\file_exists($pathName->getUploadImagePath() . $productImage->getImagePath())) {
+            \unlink($pathName->getUploadImagePath() . $productImage->getImagePath());
+        }
 
         return $this->redirectToRoute('profile_edit', [
             'id' => $productImage->getProductId()
         ]);
-
     }
-
-    public function contact(Request $request)
-    {
-
-        $product = $this->getDoctrine()
-            ->getRepository(Product::class)
-            ->previewFindById(13)
-        ;
-
-        $productInfo = \array_shift($product);
-        $productImages = $product;
-/*
-        echo '<pre>';
-        return new Response(var_dump($productImages));
-        echo '</pre>';
-        die;
-*/
-      //  $defaultData = array('message' => 'Type your message here');
-        $form = $this->createFormBuilder($productImages)
-            ->add('name', TextType::class)
-            ->add('email', EmailType::class)
-            ->add('message', TextareaType::class);
-        foreach ($productImages as $productImage){
-            $form->add('photos', CheckboxType::class, [
-                'data' => $productImage->getId()
-            ]);
-        }
-
-
-        $form
-            ->add('send', SubmitType::class)
-            ->getForm();
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            // data is an array with "name", "email", and "message" keys
-            $data = $form->getData();
-            return new Response(var_dump($data));
-        }
-
-        // ... render the form
-        return $this->render('profile/contact.html.twig', [
-            'editProduct' => $form->createView(),
-        ]);
-    }
-
 }
